niffler-python-tests/requirements.txt
selene>=2.0.0rc7
selenium
pytest
requests
python-dotenv
pydantic
faker
psycopg2
sqlalchemy
sqlmodel
allure-pytest
requests_toolbelt